+++ canonical_domain/apps.py
@@ -1,19 +1,26 @@
 from django.apps import AppConfig
 from django.conf import settings
-from django.core.checks import Error, Tags, register
+from django.core.checks import Error, Tags, Warning, register
 
 
 class CanonicalDomainAppConfig(AppConfig):
     name = "canonical_domain"
 
     def ready(self):
-        register(Tags.security)(check)
+        register(Tags.security)(default_checks)
+        register(Tags.security, deploy=True)(deploy_checks)
 
 
-def check(**kwargs):
+def default_checks(**kwargs):
     errors = []
     errors.extend(_check_ordering())
     errors.extend(_check_removed_settings())
+    return errors
+
+
+def deploy_checks(**kwargs):
+    errors = []
+    errors.extend(_check_deployment_settings())
     return errors
 
 
@@ -48,4 +55,17 @@ def _check_removed_settings():
             "You are using the removed CANONICAL_DOMAIN_SECURE setting.",
             hint="Replace CANONICAL_DOMAIN_SECURE with SECURE_SSL_REDIRECT.",
             id="canonical_domain.E003",
+        )
+
+
+def _check_deployment_settings():
+    if not settings.SECURE_SSL_HOST:
+        yield Warning(
+            "SECURE_SSL_HOST is not set.",
+            id="canonical_domain.W004",
+        )
+    if not settings.SECURE_SSL_REDIRECT:
+        yield Warning(
+            "SECURE_SSL_REDIRECT is not set to True.",
+            id="canonical_domain.W005",
         )